# collection.openapi.yaml
openapi: 3.0.3
info:
  title: FlyNext
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost
paths:
...
  /api/book/checkout:
    post:
      tags:
        - Kenson
      summary: Checkout
      description: Checkout
      operationId: checkout
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cardNumber:
                  type: string
                  example: '4111111111111111'
                cvv:
                  type: string
                  example: '123'
                expiryDate:
                  type: string
                  example: 12/25
                itineraryId:
                  type: string
                  example: itinerary_1
                totalAmount:
                  type: number
                  example: 100
                userId:
                  type: string
                  example: customer_1
            examples:
              Checkout:
                value:
                  cardNumber: '4111111111111111'
                  cvv: '123'
                  expiryDate: 12/25
                  itineraryId: itinerary_1
                  totalAmount: 100
                  userId: customer_1
      responses:
        '200':
          description: ''
...
  /api/hotel_search/search:
    get:
      tags:
        - William
      summary: >-
        search hotel by check-in date, check-out date, and city, and no
        additional filter
      description: >-
        search hotel by check-in date, check-out date, and city, and no
        additional filter
      operationId: searchHotelByCheckInDateCheckOutDateAndCityAndNoAdditionalFilter
      parameters:
        - name: city
          in: query
          schema:
            type: string
            example: Toronto
        - name: checkInDate
          in: query
          schema:
            type: string
            example: '2025-03-10'
        - name: checkOutDate
          in: query
          schema:
            type: string
            example: '2025-03-15'
        - name: name
          in: query
          schema:
            type: string
            example: Grand%20Hotel
        - name: starRating
          in: query
          schema:
            type: string
            example: '5'
        - name: minPrice
          in: query
          schema:
            type: string
            example: '100'
        - name: maxPrice
          in: query
          schema:
            type: string
            example: '250'
      responses:
        '200':
          description: ''
      security:
        - {}
...
components:
  securitySchemes:
    jwtBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
...